Break out initialization stuff
